refactor(dock): extract context menu
